move some functions from core to util
